requirements.txt
discord.py
python-dotenv
Pillow
pyyaml